feat: takes target API into account when generating .csproj
